+++ components/style/properties/helpers/animated_properties.mako.rs
@@ -23,14 +23,11 @@ use std::ptr;
 use std::mem::{self, ManuallyDrop};
 use crate::hash::FxHashMap;
 use super::ComputedValues;
-use crate::values::CSSFloat;
 use crate::values::animated::{Animate, Procedure, ToAnimatedValue, ToAnimatedZero};
 use crate::values::animated::effects::Filter as AnimatedFilter;
-use crate::values::animated::transform::{Perspective, Scale3D, Translate3D};
 #[cfg(feature = "gecko")] use crate::values::computed::TransitionProperty;
 use crate::values::computed::{ClipRect, Context};
 use crate::values::computed::ToComputedValue;
-use crate::values::computed::transform::Matrix3D;
 use crate::values::distance::{ComputeSquaredDistance, SquaredDistance};
 use crate::values::generics::effects::Filter;
 use void::{self, Void};
@@ -826,226 +823,6 @@ impl Animate for ClipRect {
 impl ToAnimatedZero for ClipRect {
     #[inline]
     fn to_animated_zero(&self) -> Result<Self, ()> { Err(()) }
-}
-
-// We implement these methods for interpolation between two Matrix3D.
-// The interpolation code is in style/values/animated/transform.rs.
-impl Matrix3D {
-    /// Return true if this has 3D components.
-    pub fn is_3d(&self) -> bool {
-        self.m13 != 0.0 || self.m14 != 0.0 ||
-        self.m23 != 0.0 || self.m24 != 0.0 ||
-        self.m31 != 0.0 || self.m32 != 0.0 || self.m33 != 1.0 || self.m34 != 0.0 ||
-        self.m43 != 0.0 || self.m44 != 1.0
-    }
-
-    /// Return determinant value.
-    pub fn determinant(&self) -> CSSFloat {
-        self.m14 * self.m23 * self.m32 * self.m41 -
-        self.m13 * self.m24 * self.m32 * self.m41 -
-        self.m14 * self.m22 * self.m33 * self.m41 +
-        self.m12 * self.m24 * self.m33 * self.m41 +
-        self.m13 * self.m22 * self.m34 * self.m41 -
-        self.m12 * self.m23 * self.m34 * self.m41 -
-        self.m14 * self.m23 * self.m31 * self.m42 +
-        self.m13 * self.m24 * self.m31 * self.m42 +
-        self.m14 * self.m21 * self.m33 * self.m42 -
-        self.m11 * self.m24 * self.m33 * self.m42 -
-        self.m13 * self.m21 * self.m34 * self.m42 +
-        self.m11 * self.m23 * self.m34 * self.m42 +
-        self.m14 * self.m22 * self.m31 * self.m43 -
-        self.m12 * self.m24 * self.m31 * self.m43 -
-        self.m14 * self.m21 * self.m32 * self.m43 +
-        self.m11 * self.m24 * self.m32 * self.m43 +
-        self.m12 * self.m21 * self.m34 * self.m43 -
-        self.m11 * self.m22 * self.m34 * self.m43 -
-        self.m13 * self.m22 * self.m31 * self.m44 +
-        self.m12 * self.m23 * self.m31 * self.m44 +
-        self.m13 * self.m21 * self.m32 * self.m44 -
-        self.m11 * self.m23 * self.m32 * self.m44 -
-        self.m12 * self.m21 * self.m33 * self.m44 +
-        self.m11 * self.m22 * self.m33 * self.m44
-    }
-
-    /// Transpose a matrix.
-    pub fn transpose(&self) -> Self {
-        Self {
-            % for i in range(1, 5):
-            % for j in range(1, 5):
-            m${i}${j}: self.m${j}${i},
-            % endfor
-            % endfor
-        }
-    }
-
-    /// Return inverse matrix.
-    pub fn inverse(&self) -> Result<Matrix3D, ()> {
-        let mut det = self.determinant();
-
-        if det == 0.0 {
-            return Err(());
-        }
-
-        det = 1.0 / det;
-        let x = Matrix3D {
-            m11: det *
-            (self.m23*self.m34*self.m42 - self.m24*self.m33*self.m42 +
-             self.m24*self.m32*self.m43 - self.m22*self.m34*self.m43 -
-             self.m23*self.m32*self.m44 + self.m22*self.m33*self.m44),
-            m12: det *
-            (self.m14*self.m33*self.m42 - self.m13*self.m34*self.m42 -
-             self.m14*self.m32*self.m43 + self.m12*self.m34*self.m43 +
-             self.m13*self.m32*self.m44 - self.m12*self.m33*self.m44),
-            m13: det *
-            (self.m13*self.m24*self.m42 - self.m14*self.m23*self.m42 +
-             self.m14*self.m22*self.m43 - self.m12*self.m24*self.m43 -
-             self.m13*self.m22*self.m44 + self.m12*self.m23*self.m44),
-            m14: det *
-            (self.m14*self.m23*self.m32 - self.m13*self.m24*self.m32 -
-             self.m14*self.m22*self.m33 + self.m12*self.m24*self.m33 +
-             self.m13*self.m22*self.m34 - self.m12*self.m23*self.m34),
-            m21: det *
-            (self.m24*self.m33*self.m41 - self.m23*self.m34*self.m41 -
-             self.m24*self.m31*self.m43 + self.m21*self.m34*self.m43 +
-             self.m23*self.m31*self.m44 - self.m21*self.m33*self.m44),
-            m22: det *
-            (self.m13*self.m34*self.m41 - self.m14*self.m33*self.m41 +
-             self.m14*self.m31*self.m43 - self.m11*self.m34*self.m43 -
-             self.m13*self.m31*self.m44 + self.m11*self.m33*self.m44),
-            m23: det *
-            (self.m14*self.m23*self.m41 - self.m13*self.m24*self.m41 -
-             self.m14*self.m21*self.m43 + self.m11*self.m24*self.m43 +
-             self.m13*self.m21*self.m44 - self.m11*self.m23*self.m44),
-            m24: det *
-            (self.m13*self.m24*self.m31 - self.m14*self.m23*self.m31 +
-             self.m14*self.m21*self.m33 - self.m11*self.m24*self.m33 -
-             self.m13*self.m21*self.m34 + self.m11*self.m23*self.m34),
-            m31: det *
-            (self.m22*self.m34*self.m41 - self.m24*self.m32*self.m41 +
-             self.m24*self.m31*self.m42 - self.m21*self.m34*self.m42 -
-             self.m22*self.m31*self.m44 + self.m21*self.m32*self.m44),
-            m32: det *
-            (self.m14*self.m32*self.m41 - self.m12*self.m34*self.m41 -
-             self.m14*self.m31*self.m42 + self.m11*self.m34*self.m42 +
-             self.m12*self.m31*self.m44 - self.m11*self.m32*self.m44),
-            m33: det *
-            (self.m12*self.m24*self.m41 - self.m14*self.m22*self.m41 +
-             self.m14*self.m21*self.m42 - self.m11*self.m24*self.m42 -
-             self.m12*self.m21*self.m44 + self.m11*self.m22*self.m44),
-            m34: det *
-            (self.m14*self.m22*self.m31 - self.m12*self.m24*self.m31 -
-             self.m14*self.m21*self.m32 + self.m11*self.m24*self.m32 +
-             self.m12*self.m21*self.m34 - self.m11*self.m22*self.m34),
-            m41: det *
-            (self.m23*self.m32*self.m41 - self.m22*self.m33*self.m41 -
-             self.m23*self.m31*self.m42 + self.m21*self.m33*self.m42 +
-             self.m22*self.m31*self.m43 - self.m21*self.m32*self.m43),
-            m42: det *
-            (self.m12*self.m33*self.m41 - self.m13*self.m32*self.m41 +
-             self.m13*self.m31*self.m42 - self.m11*self.m33*self.m42 -
-             self.m12*self.m31*self.m43 + self.m11*self.m32*self.m43),
-            m43: det *
-            (self.m13*self.m22*self.m41 - self.m12*self.m23*self.m41 -
-             self.m13*self.m21*self.m42 + self.m11*self.m23*self.m42 +
-             self.m12*self.m21*self.m43 - self.m11*self.m22*self.m43),
-            m44: det *
-            (self.m12*self.m23*self.m31 - self.m13*self.m22*self.m31 +
-             self.m13*self.m21*self.m32 - self.m11*self.m23*self.m32 -
-             self.m12*self.m21*self.m33 + self.m11*self.m22*self.m33),
-        };
-
-        Ok(x)
-    }
-
-    /// Multiplies `pin * self`.
-    pub fn pre_mul_point4(&self, pin: &[f32; 4]) -> [f32; 4] {
-        [
-        % for i in range(1, 5):
-            pin[0] * self.m1${i} +
-            pin[1] * self.m2${i} +
-            pin[2] * self.m3${i} +
-            pin[3] * self.m4${i},
-        % endfor
-        ]
-    }
-
-    /// Multiplication of two 4x4 matrices.
-    pub fn multiply(&self, other: &Self) -> Self {
-        Matrix3D {
-        % for i in range(1, 5):
-            % for j in range(1, 5):
-                m${i}${j}:
-                    self.m${i}1 * other.m1${j} +
-                    self.m${i}2 * other.m2${j} +
-                    self.m${i}3 * other.m3${j} +
-                    self.m${i}4 * other.m4${j},
-            % endfor
-        % endfor
-        }
-    }
-
-    /// Scale the matrix by a factor.
-    #[inline]
-    pub fn scale_by_factor(&mut self, scaling_factor: CSSFloat) {
-        % for i in range(1, 5):
-        % for j in range(1, 5):
-            self.m${i}${j} *= scaling_factor;
-        % endfor
-        % endfor
-    }
-
-    /// This is used by retrieving the scale and shear factors
-    /// during decomposing a 3d matrix.
-    #[inline]
-    pub fn get_matrix_3x3_part(&self) -> [[f32; 3]; 3] {
-        let mut row: [[f32; 3]; 3] = [[0.0; 3]; 3];
-        % for i in range(1, 4):
-            row[${i - 1}][0] = self.m${i}1;
-            row[${i - 1}][1] = self.m${i}2;
-            row[${i - 1}][2] = self.m${i}3;
-        % endfor
-        row
-    }
-
-    /// Set perspective on the matrix.
-    #[inline]
-    pub fn set_perspective(&mut self, perspective: &Perspective) {
-        % for i in range(1, 5):
-            self.m${i}4 = perspective.${i - 1};
-        % endfor
-    }
-
-    /// Apply translate on the matrix.
-    #[inline]
-    pub fn apply_translate(&mut self, translate: &Translate3D) {
-        % for i in range(1, 5):
-        % for j in range(1, 4):
-            self.m4${i} += translate.${j - 1} * self.m${j}${i};
-        % endfor
-        % endfor
-    }
-
-    /// Apply scale on the matrix
-    #[inline]
-    pub fn apply_scale(&mut self, scale: &Scale3D) {
-        % for i in range(1, 4):
-        % for j in range(1, 5):
-            self.m${i}${j} *= scale.${i - 1};
-        % endfor
-        % endfor
-    }
-
-    /// Negate the matrix and the scaling factors.
-    /// This is a helper function for interpolation the Matrix.
-    #[inline]
-    pub fn negate_matrix_3x3_and_scaling_factor(row: &mut [[f32; 3]; 3], scale: &mut Scale3D) {
-        % for i in range(3):
-            scale.${i} *= -1.0;
-            row[${i}][0] *= -1.0;
-            row[${i}][1] *= -1.0;
-            row[${i}][2] *= -1.0;
-        % endfor
-    }
 }
 
 <%
